Fix cloudinary file upload issue

+++ src/utils/cloudinary.js
@@ -1,5 +1,5 @@
 import { v2 as cloudinary } from "cloudinary";
-import fs from "fs"
+import fs from "fs";
 
 
 cloudinary.config({ 
@@ -16,8 +16,7 @@ const uploadOnCloudinary = async (localFilePath)=>{
             resource_type: "auto"
         })
         // file uploaded successfully
-        // console.log("File has been uploaded successfully on cloudinary", response.url)
-        fs.unlinkSync(localFilePath)
+        if(response) fs.unlinkSync(localFilePath);
         return response;
     } catch (error) {
         fs.unlinkSync(localFilePath) //remove the locally saved temporary file as the upload operation failed
